+++ index.js
@@ -13,24 +13,33 @@ const _detectArgs = require("./src/functions/detectArgs");
 
 const Verbose = require("./src/classes/Verbose");
 
-Verbose.custom("Initialized arguments.");
+Verbose.custom("Detecting no timebomb argument...");
+if (!_detectArgs("timebomb")) require("./src/functions/init/timebomb")();
 
-if (_detectArgs("timebomb")) require("./src/functions/init/timebomb")();
-
+Verbose.custom("Detecting no checks argument...");
 if (_detectArgs("checks")) require("./src/functions/init/startupChecks")();
 
+Verbose.custom("Detecting startup arguments help argument...");
 if (_detectArgs("help")) {
   require("./src/functions/init/startupArgs")();
+
+  Verbose.exitProcess();
   process.exit(0);
 }
 
+Verbose.custom("Detecting version argument...");
 if (_detectArgs("version")) {
+  Verbose.custom("Showing version information...");
   console.log(`${GLOBAL_NAME}, v${VERSION} (build ${BUILD})\n`);
+
+  Verbose.exitProcess();
   process.exit(0);
 }
 
+Verbose.custom("Starting pre-boot interpreter...");
 require("./src/functions/init/preBootInt")();
 
+Verbose.custom("Initializing arguments...");
 require("./src/functions/init/initArgs");
 
 // Repeat forever until the user exits