+++ src/functions/init/initArgs.js
@@ -6,6 +6,7 @@ const Verbose = require("../../classes/Verbose");
 
 const { GLOBAL_NAME, EXPIRY_DATE } = require("../../variables/constants");
 
+Verbose.custom("Creating configuration manager...");
 const config = new ConfigManager();
 
 Verbose.custom("Displaying intro...");
@@ -13,7 +14,7 @@ if (!_detectArgs("intro")) require("./intro");
 
 Verbose.custom("Completing timebomb disabler check...");
 if (
-  !_detectArgs("timebomb") &&
+  _detectArgs("timebomb") &&
   !_detectArgs("warnings") &&
   EXPIRY_DATE.getTime() < new Date().getTime()
 )
@@ -21,17 +22,12 @@ if (
     `The timebomb has been disabled. The timebomb is a security feature to prevent you from using outdated software. Please upgrade to a new version of ${GLOBAL_NAME} as soon as possible.`
   );
 
-// TODO Remove this when verbose mode is fully implemented
-if (_detectArgs("verbose"))
-  InfoMessages.info(
-    `The verbose setting for ${GLOBAL_NAME} will be introduced in the next build. It is currently unused.`
-  );
-
 Verbose.custom("Completing fatal error file dump check...");
 if (_detectArgs("dump") && !_detectArgs("warnings"))
   InfoMessages.warning("The fatal error file dump feature has been disabled.");
 
 if (_detectArgs("reset")) {
+  Verbose.custom("Deleting configuration file...");
   if (config.deleteConfig()) {
     InfoMessages.success(`Successfully deleted the ${GLOBAL_NAME} configuration file.`);
   } else {
@@ -41,6 +37,7 @@ if (_detectArgs("reset")) {
   }
 }
 
+Verbose.custom("Creating configuration file...");
 if (!config.createConfig())
   InfoMessages.error(
     `${GLOBAL_NAME} failed to create the configuration file. Some features will not work.`